split tests into dedicated classes
